# todo-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Todo API",
        "description": "This is the API for managing Todo tasks",
        "version": "1.0.0"
    },
    "paths": {
        "/api/todos": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Get all Todos",
                "operationId": "521da6d629de5b14a88d7b8c218cb6a1",
                "responses": {
                    "200": {
                        "description": "A list of todos"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Create a new Todo",
                "operationId": "9b07ee37ef539d3573ba1ca6c3ad7fb1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Todo Task"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description of the task"
                                    },
                                    "completed": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Todo created successfully"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/todos/{id}": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Get a Todo by ID",
                "operationId": "c7cfad681d3a421509390996e9c30925",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Todo to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo found"
                    },
                    "404": {
                        "description": "Todo not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Todos"
                ],
                "summary": "Update an existing Todo",
                "operationId": "3b87e27d1714d5f0f389613c9f54bea2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Todo Task"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description of the task"
                                    },
                                    "completed": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Todo updated successfully"
                    },
                    "404": {
                        "description": "Todo not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Todos"
                ],
                "summary": "Delete an existing Todo",
                "operationId": "3db9089ef72b760f725785301a469fe7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Todo deleted successfully"
                    },
                    "404": {
                        "description": "Todo not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Todos",
            "description": "Todos"
        }
    ]
}